cleanup_duplication: moved duplications to abstract class.
